Organize all the tests in to a test folder. Updated paths to components in import statement in the test. Also update key name for reason for vist, according to the recent update on the API.
